docs: README: specify how to contribute instead of pointing at another file
This repository is partially included in another repository: poky.

However its README isn't making it, so documentation/README pointing at
the README at the root of the git repository would lead the contributor
nowhere.

Instead, let's include the appropriate information directly in
documentation/README which does make it to the poky git repo.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>