deduplicate README
Having a README at the root of the repo is a good practice so we should
keep it. However, it contains information that is now duplicated in
documentation/README. Additionally, this file isn't included in poky
while what's under documentation/ directory is.

Avoid having to maintain both by simply making the root README into a
symlink to documentation/README.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>