ref-manual/yocto-project-supported-features: add ppc maintainers
PowerPC (32-bit) remains relevant for our use cases, and we aim to
maintain support as long as it is feasible with reasonable effort.

For transparency, our primary focus is on maintaining
core-image-minimal. Support for more complex images may be limited
unless additional contributors join, as ongoing testing and maintenance
of these images may not be feasible otherwise.

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>