poky.yaml.in: increase required disk space to 140G
Increase the minimum size requirement to 140G, after running a build
with distro Poky. The size after rm_work has increase of only 1G, so
leave the current value for MIN_DISK_SPACE_RM_WORK.

Also, update the comment to mention that this is for distro Poky. It was
mentioned in commit message 285ba689ce3a ("manuals: update disk space
requirements"), but not in the comment here.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>