Makefile: add a checks rule
The checks rule can be run to perform some checks on the docs.
For now it only includes running the check-glossaries script.

This rule should only run basic checks or should not prevent the user
from building the docs in case of failure.

This rule can be used in the future to run additional checks on the
documentation.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>