dev-manual: add a hash equivalence server setup document
Add a small document to describe how to start a Hash Equivalence server
with bitbake-hashserv, the reference server provided by BitBake.

Update reference to hash equivalence in other places in the
documentation to mention this document.

Move the note from the concepts document to the new document, as a
warning note.

[YOCTO #15921]

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>