brief-yoctoprojectqs: specify what "recent Ubuntu Linux distribution" is
The docs from any release can be read at any moment. The time-specific
"recent" doesn't actually help. The user could be reading the
instructions for kirkstone and try to build on Ubuntu 25.10 as that is
the most recent version. It'll likely fail though.

Instead, point at the list of supported Linux distributions so that they
know which are supported for the version of the docs they are currently
reading.

Reported-by: Robert P. J. Day <rpjday@crashcourse.ca>
Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>