make sure Quick Build section and System Requirements are in sync
Right now, there are additional instructions that are missing in the
Quick Build section compared to the ones in System Requirements.

By including the exact same rST file for Ubuntu/Debian in both sections,
we can make sure both are kept in sync while still providing only
instructions for Ubuntu and not requiring the user to visit another
page.

Signed-off-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>