overview-manual/concepts.rst: update the cross-development toolchain section
Update the cross-development toolchain section of the concepts document
to make it GCC-agnostic as we now have Clang support.

The chain presented to build the relocatable toolchain
(gcc-cross-canadian) was also updated and simplified. `bitbake -g
gcc-cross-canadian-aarch64` was used to understand it better.

Also move the gcc-cross-canadian note after introducing the relocatable
chain of dependencies, as it felt a little too soon to mention it at
this point.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>