brief-yoctoprojectqs/index.rst: update available bitbake-setup configurations
Configurations were updated with commit 720df1a53452 ("bitbake-setup:
add whinlatter configs for poky and oe-nodistro") in BitBake. To more
accurately represent what the output will be in the real world, add them
here by using the DISTRO_NAME_NO_CAP and DISTRO_RELEASE_SERIES meta
variable that are translated to "whinlatter" and "5.3" respectively.

Note that the real output mentions "(supported until 2026-05-31)" but we
don't want to maintain that value over time, so just omit it.

Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>