contributor-guide/submit-changes.rst: remove mention of Upstream-Status
The Upstream-Status tag is not used for patches sent on mailing lists,
but for patches to be applied to recipes. To avoid confusion in this
Contributor Guide, remove the mention of this tag as part of the
Patchtest examples as it can be misinterpreted as something to include
in a regular patch.

Reviewed-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>