overview-manual/yp-intro.rst: remove obsolete poky repo references
Remove references to the Poky repository. This document had a lot of
reference to it, so a lot was removed. The "Reference Distribution
(Poky)" section was removed as there already was a "Reference Embedded
Distribution (Poky)" below. This section was simplified as now Poky is
just a distro, not a repo. The historical section was also updated to
mention the obsolete Poky repo.

Reviewed-by: Quentin Schulz <quentin.schulz@cherry.de>
Signed-off-by: Antonin Godard <antonin.godard@bootlin.com>